+++ mc12111/src/main/java/com/mcvoice/ttvoice/screen/AdvancedConfigScreen.java
@@ -30,7 +30,7 @@ public class AdvancedConfigScreen extends Screen {
         int x = centerX - buttonWidth / 2;
         int y = 38 - scrollY;
 
-        addRenderableWidget(new StringWidget(centerX - buttonWidth / 2, 14, buttonWidth, 20,
+        addRenderableWidget(new StringWidget(centerX - buttonWidth / 2, 14 - scrollY, buttonWidth, 20,
             Component.translatable("config.mcvoice.advanced"), font));
 
         AbstractSliderButton volumeSlider = new AbstractSliderButton(